+++ editor/renderer/src/__tests__/vr-tests/text-highlighter.vr.ts
@@ -1,4 +1,9 @@
 import { snapshot } from '@af/visual-regression';
-import { RendererWithTextHighlighter } from '../__helpers/renderer-with-text-highlighter';
+import {
+  RendererWithTextHighlighter,
+  RendererWithFilteredTextHighlighter,
+} from '../__helpers/renderer-with-text-highlighter';
 
 snapshot(RendererWithTextHighlighter);
+
+snapshot(RendererWithFilteredTextHighlighter);